start implementing SSL_CONF_* APIs
This commit introduces the scaffolding for the `SSL_CONF_*` APIs that
were previously stubbed. All of the API fns with the exception of
`SSL_CONF_cmd_argv` are now implemented as of this commit.

Note: we do not yet suppor the vast majority of possible `SSL_CONF_cmd`
options. To start, only `MinProtocol` and `MaxProtocol` are supported.
Additional commands will be supported in subsequent commits.