implement SSL_get_negotiated_group
Rustls 0.23.11 added a `negotiated_key_exchange_group()` to connections
that exposes the `&dyn SupportedKxGroup` used by the connection. This
allows us to implement `SSL_get_negotiated_group()`.

Our implementation differs slightly from the upstream w.r.t what NID we
return for resumed connections. Rustls will return `None` (which we
translate to `NID_undef`) for connections that are still handshaking, or
for resumed TLS 1.2 handshakes. In contrast, OpenSSL returns the
negotiated group of the original connection in that latter case.